requirements.txt
# requirements.txt
pandas>=1.0.0
openai>=0.27.0
streamlit>=1.0.0
python-dotenv>=0.21.0
